admin validity check method

+++ AlifMidTermProject/Program.cs
@@ -10,29 +10,39 @@ namespace AlifMidTermProject
             {
                 case "1":
                     {
-                        Console.WriteLine(@"Please choose action: 
+                        if (AdminCredentials.adminValidityCheck() == true)
+                        {
+                            Console.WriteLine("Welcome!");
+                            Console.WriteLine(@"Please choose action: 
                                  1. Create new Admin
                                  2. List of admins
                                  3. Delete admin by id
                                  ");
-                        int action = Convert.ToInt32(Console.ReadLine());
-                        switch (action)
-                        {
-                            case 1:
-                                AdminCredentials.insertAdmin();
-                                break;
-                            case 2:
-                                AdminCredentials.selectAllAdmin();
-                                break;
-                            case 3:
-                                AdminCredentials.deleteByIdAdmin();
-                                break;
-
+                            int action = Convert.ToInt32(Console.ReadLine());
+                            switch (action)
+                            {
+                                case 1:
+                                    AdminCredentials.insertAdmin();
+                                    break;
+                                case 2:
+                                    AdminCredentials.selectAllAdmin();
+                                    break;
+                                case 3:
+                                    AdminCredentials.deleteByIdAdmin();
+                                    break;
+                            }
+                            break;
                         }
-
-                        break;
-
+                        else
+                        {
+                            Console.WriteLine("Invalid data!");
+                        }                       
                     }
+                    break;
+                case "2":
+                    
+                    break;
+
 
             }
         }
